refactor(api): extract health/version routes (phase 1 of #793)
Move GET /api/version and GET /api/health out of the create_app()
closure in maxwell_daemon/api/server.py into a new
maxwell_daemon/api/routes/health.py module with a register(app, daemon)
entry point.  This is the first phase of decomposing the 3,300+ line
god object tracked in #793; subsequent phases will extract status,
tasks, dispatch, control, fleet, websocket, and middleware groups.

Response shapes (HealthResponse, VersionResponse) are unchanged — the
append-only HTTP contract is preserved.  All existing tests in
tests/unit/test_api.py and tests/unit/test_api_contract.py pass
without modification.

Partial #793
